refactor(structure): typo en structure
