+++ src/main/java/com/mgcss/mgcss_track_7/infrastructure/persistence/JpaClienteRepositorio.java
@@ -1,4 +1,4 @@
-package com.mgcss.mgcss_track_7.infraestrucure.persistence;
+package com.mgcss.mgcss_track_7.infrastructure.persistence;
 
 import org.springframework.data.jpa.repository.JpaRepository;
