+++ src/main/java/com/mgcss/mgcss_track_7/infrastructure/persistence/TecnicoEntidad.java
@@ -1,4 +1,4 @@
-package com.mgcss.mgcss_track_7.infraestrucure.persistence;
+package com.mgcss.mgcss_track_7.infrastructure.persistence;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
