+++ src/main/java/com/mgcss/mgcss_track_7/service/ServicioSolicitud.java
@@ -5,8 +5,9 @@ import java.util.Optional;
 import com.mgcss.mgcss_track_7.domain.Cliente;
 import com.mgcss.mgcss_track_7.domain.Solicitud;
 import com.mgcss.mgcss_track_7.domain.Tecnico;
+import com.mgcss.mgcss_track_7.infrastructure.persistence.SolicitudRepositorio;
+
 import org.springframework.stereotype.Service;
-import com.mgcss.mgcss_track_7.infraestrucure.persistence.SolicitudRepositorio;
 
 import java.util.List;
