+++ src/test/java/com/mgcss/mgcss_track_7/integration/JpaClienteRepositorioTest.java
@@ -4,9 +4,10 @@ import org.springframework.boot.data.jpa.test.autoconfigure.DataJpaTest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ActiveProfiles;
 
-import com.mgcss.mgcss_track_7.infraestrucure.persistence.JpaClienteRepositorio;
-import com.mgcss.mgcss_track_7.infraestrucure.persistence.ClienteEntidad;
 import com.mgcss.mgcss_track_7.domain.Cliente;
+import com.mgcss.mgcss_track_7.infrastructure.persistence.ClienteEntidad;
+import com.mgcss.mgcss_track_7.infrastructure.persistence.JpaClienteRepositorio;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.Optional;
