+++ src/test/java/com/mgcss/mgcss_track_7/integration/JpaSolicitudRepositorioTest.java
@@ -5,13 +5,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ActiveProfiles;
 
 import com.mgcss.mgcss_track_7.domain.Cliente;
-import com.mgcss.mgcss_track_7.infraestrucure.persistence.ClienteEntidad;
-import com.mgcss.mgcss_track_7.infraestrucure.persistence.JpaClienteRepositorio;
-import com.mgcss.mgcss_track_7.infraestrucure.persistence.JpaSolicitudRepositorio;
-import com.mgcss.mgcss_track_7.infraestrucure.persistence.SolicitudEntidad;
-
 import com.mgcss.mgcss_track_7.domain.Solicitud;
-
+import com.mgcss.mgcss_track_7.infrastructure.persistence.ClienteEntidad;
+import com.mgcss.mgcss_track_7.infrastructure.persistence.JpaClienteRepositorio;
+import com.mgcss.mgcss_track_7.infrastructure.persistence.JpaSolicitudRepositorio;
+import com.mgcss.mgcss_track_7.infrastructure.persistence.SolicitudEntidad;
 
 import static org.junit.jupiter.api.Assertions.*;
