+++ src/test/java/com/mgcss/mgcss_track_7/unit/persistence/SolicitudRepositorioImplTest.java
@@ -13,11 +13,11 @@ import org.mockito.Mockito;
 import com.mgcss.mgcss_track_7.domain.Cliente;
 import com.mgcss.mgcss_track_7.domain.Solicitud;
 import com.mgcss.mgcss_track_7.domain.Tecnico;
-import com.mgcss.mgcss_track_7.infraestrucure.persistence.ClienteEntidad;
-import com.mgcss.mgcss_track_7.infraestrucure.persistence.JpaSolicitudRepositorio;
-import com.mgcss.mgcss_track_7.infraestrucure.persistence.SolicitudEntidad;
-import com.mgcss.mgcss_track_7.infraestrucure.persistence.SolicitudRepositorioImpl;
-import com.mgcss.mgcss_track_7.infraestrucure.persistence.TecnicoEntidad;
+import com.mgcss.mgcss_track_7.infrastructure.persistence.ClienteEntidad;
+import com.mgcss.mgcss_track_7.infrastructure.persistence.JpaSolicitudRepositorio;
+import com.mgcss.mgcss_track_7.infrastructure.persistence.SolicitudEntidad;
+import com.mgcss.mgcss_track_7.infrastructure.persistence.SolicitudRepositorioImpl;
+import com.mgcss.mgcss_track_7.infrastructure.persistence.TecnicoEntidad;
 
 class SolicitudRepositorioImplTest {
