+++ src/test/java/com/mgcss/mgcss_track_7/unit/service/ServicioClienteTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
 import com.mgcss.mgcss_track_7.domain.Cliente;
-import com.mgcss.mgcss_track_7.infraestrucure.persistence.ClienteRepositorio;
+import com.mgcss.mgcss_track_7.infrastructure.persistence.ClienteRepositorio;
 import com.mgcss.mgcss_track_7.service.ServicioCliente;
 
 class ServicioClienteTest {
